cw2/requirements.txt
Flask==0.10.1
Flask-WTF==0.12
git-remote-helpers==0.1.0
gspread==0.2.5
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
py-bcrypt==0.4
requests==2.7.0
Werkzeug==0.10.4
WTForms==2.0.2
